Refactor Labels util
