+++ src/nl/hannahsten/texifyidea/highlighting/LatexAnnotator.kt
@@ -9,6 +9,7 @@ import com.intellij.psi.PsiElement
 import nl.hannahsten.texifyidea.lang.Environment
 import nl.hannahsten.texifyidea.psi.*
 import nl.hannahsten.texifyidea.util.*
+import nl.hannahsten.texifyidea.util.labels.getLabelDefinitionCommands
 import nl.hannahsten.texifyidea.util.magic.CommandMagic
 
 /**
