+++ src/nl/hannahsten/texifyidea/intentions/LatexAddLabelIntention.kt
@@ -13,10 +13,7 @@ import com.intellij.refactoring.rename.RenameProcessor
 import com.intellij.refactoring.rename.inplace.MemberInplaceRenamer
 import nl.hannahsten.texifyidea.psi.LatexCommandWithParams
 import nl.hannahsten.texifyidea.psi.LatexPsiHelper
-import nl.hannahsten.texifyidea.util.endOffset
-import nl.hannahsten.texifyidea.util.findLatexAndBibtexLabelStringsInFileSet
-import nl.hannahsten.texifyidea.util.parentOfType
-import nl.hannahsten.texifyidea.util.findLatexAndBibtexLabelStringsInFileSet
+import nl.hannahsten.texifyidea.util.labels.findLatexAndBibtexLabelStringsInFileSet
 import kotlin.math.max
 
 abstract class LatexAddLabelIntention : TexifyIntentionBase("Add label") {
