+++ src/nl/hannahsten/texifyidea/navigation/GotoLabelSymbolContributor.kt
@@ -2,8 +2,8 @@ package nl.hannahsten.texifyidea.navigation
 
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
-import nl.hannahsten.texifyidea.util.extractLabelName
-import nl.hannahsten.texifyidea.util.findAllLabelsAndBibtexIds
+import nl.hannahsten.texifyidea.util.labels.extractLabelName
+import nl.hannahsten.texifyidea.util.labels.findAllLabelsAndBibtexIds
 
 /**
  * @author Hannah Schellekens
