+++ src/nl/hannahsten/texifyidea/util/labels/ProjectLabels.kt
@@ -1,0 +1,54 @@
+package nl.hannahsten.texifyidea.util.labels
+
+import com.intellij.openapi.project.Project
+import com.intellij.psi.PsiElement
+import nl.hannahsten.texifyidea.index.BibtexEntryIndex
+import nl.hannahsten.texifyidea.index.LatexCommandsIndex
+import nl.hannahsten.texifyidea.index.LatexParameterLabeledCommandsIndex
+import nl.hannahsten.texifyidea.index.LatexParameterLabeledEnvironmentsIndex
+import nl.hannahsten.texifyidea.lang.CommandManager
+import nl.hannahsten.texifyidea.util.magic.CommandMagic
+
+
+/**
+ * Finds all defined labels within the project, including bibtex entries.
+ *
+ * @return The found label commands.
+ */
+fun Project.findAllLabelsAndBibtexIds(): Collection<PsiElement> {
+    val commands = LatexCommandsIndex.getItems(this).findLatexCommandsLabels(this)
+    val bibtexIds = BibtexEntryIndex.getIndexedEntries(this)
+    val environments = LatexParameterLabeledEnvironmentsIndex.getItems(this)
+    val parameterLabeledCommands = LatexParameterLabeledCommandsIndex.getItems(this)
+    val result = ArrayList<PsiElement>(commands)
+    result.addAll(bibtexIds)
+    result.addAll(environments)
+    result.addAll(parameterLabeledCommands)
+    return result
+}
+
+/**
+ * All commands that represent a reference to a label, including user defined commands.
+ */
+fun Project.getLabelReferenceCommands(): Set<String> {
+    CommandManager.updateAliases(CommandMagic.labelReferenceWithoutCustomCommands, this)
+    return CommandManager.getAliases(CommandMagic.labelReferenceWithoutCustomCommands.first())
+}
+
+/**
+ * Get all commands defining labels, including user defined commands. This will not check if the aliases need to be updated.
+ */
+fun getLabelDefinitionCommands() = CommandManager.getAliases(CommandMagic.labelDefinitionsWithoutCustomCommands.first())
+
+/**
+ * Get all commands defining labels, including user defined commands.
+ * If you need to know which parameters of user defined commands define a label, use [CommandManager.labelAliasesInfo].
+ *
+ * This will check if the cache of user defined commands needs to be updated, based on the given project, and therefore may take some time.
+ */
+fun Project.getLabelDefinitionCommands(): Set<String> {
+    // Check if updates are needed
+    CommandManager.updateAliases(CommandMagic.labelDefinitionsWithoutCustomCommands, this)
+    return CommandManager.getAliases(CommandMagic.labelDefinitionsWithoutCustomCommands.first())
+}
+
