+++ test/resources/inspections/latex/unresolvedreference/xr-test.tex
@@ -11,7 +11,6 @@
 
 \begin{document}
 
-%    Slide \ref{P-slide:first} todo
-    Slide \ref{slide:first}
+    Slide~\ref{P-slide:first}
 
 \end{document}
